+++ comics/comics/yamac.py
@@ -5,17 +5,17 @@ from comics.core.comic_data import ComicDataBase
 class ComicData(ComicDataBase):
     name = 'you and me and cats'
     language = 'en'
-    url = 'http://strawberry-pie.net/yamac/'
+    url = 'http://strawberry-pie.net/SA/'
     start_date = '2009-07-01'
     rights = 'bubble'
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 90
+    history_capable_days = 365
     time_zone = 'US/Pacific'
 
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://strawberry-pie.net/yamac/?feed=rss2')
+        feed = self.parse_feed('http://strawberry-pie.net/SA/?feed=rss2')
         for entry in feed.for_date(pub_date):
             if 'comic' not in entry.tags:
                 continue