+++ comics/comics/crookedgremlins.py
@@ -11,22 +11,20 @@ class ComicData(ComicDataBase):
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 180
+    history_capable_date = '2008-04-01'
     time_zone = 'US/Pacific'
 
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://crookedgremlins.com/feed/')
-        for entry in feed.for_date(pub_date):
-            if not 'Comics' in entry.tags:
-                continue
-            title = entry.title
-            url = entry.summary.src('img[src*="/comics/"]')
+        page = self.parse_page(
+            'http://crookedgremlins.com/%s/' % pub_date.strftime('%Y/%m/%d'))
+        title = page.alt('#comic img')
+        url = page.src('#comic img')
 
-            # Put together the text from multiple paragraphs
-            text_paragraphs = entry.summary.text('p', allow_multiple=True)
-            if text_paragraphs is not None:
-                text = '\n\n'.join(text_paragraphs)
-            else:
-                text = None
+        # Put together the text from multiple paragraphs
+        text_paragraphs = page.text('.post-content p', allow_multiple=True)
+        if text_paragraphs is not None:
+            text = '\n\n'.join(text_paragraphs)
+        else:
+            text = None
 
-            return CrawlerImage(url, title, text)
+        return CrawlerImage(url, title, text)