+++ docs/changes.rst
@@ -15,6 +15,7 @@ v2.3.1 (UNRELEASED)
 - New: ``iamarg``
 - Update: ``lookingforgroup`` after feed change.
 - Update: ``phd`` after feed change.
+- Update: ``poledancingadventures`` after site change.
 - Update: ``yamac`` after site change.
 - Inactive: ``boxerhockey``
 - Inactive: ``eatthattoast``