+++ docs/changes.rst
@@ -17,6 +17,7 @@ v2.3.1 (UNRELEASED)
 - Update: ``phd`` after feed change.
 - Update: ``poledancingadventures`` after site change.
 - Update: ``pcweenies`` after feed change.
+- Update: ``playervsplayer`` after site change.
 - Update: ``yamac`` after site change.
 - Inactive: ``antics``
 - Inactive: ``boxerhockey``