+++ docs/changes.rst
@@ -13,6 +13,7 @@ v2.3.1 (UNRELEASED)
 **Crawlers**
 
 - New: ``iamarg``
+- Update: ``hijinksensue`` after feed change.
 - Update: ``lookingforgroup`` after feed change.
 - Update: ``phd`` after feed change.
 - Update: ``poledancingadventures`` after site change.